// src/features/Auth/schema.ts
import { UserRole } from "@/types/auth";
import { z } from "zod";
 
 
export const loginSchema = z.object({
  email: z.string().email({ message: "Invalid email address" }),
  password: z.string().min(6, { message: "Password must be at least 6 characters" }),
});

export const registerSchema = z.object({
  name: z.string().min(2, { message: "Name must be at least 2 characters" }),
  email: z.string().email({ message: "Invalid email address" }),
  password: z.string().min(6, { message: "Password must be at least 6 characters" }),
  role: z.nativeEnum(UserRole).default(UserRole.MEMBER),
  mosqueId: z.string().optional(), // Expected as string from form, cast to ObjectId in action
});

export type LoginInput = z.infer<typeof loginSchema>;
export type RegisterInput = z.infer<typeof registerSchema>;